FINALY FOUND OPTIMAL SWAP COMPUTING ??

+++ src/2_stack/stack.h
@@ -15,7 +15,7 @@ typedef struct s_stack
 
 bool	stack_init(t_stack *a, t_stack *b, uint *values, size_t count);
 bool	stack_dup(t_stack *dst, t_stack *src);
-uint	stack_get_value(t_stack *stack, size_t index);
+uint	stack_get_value(t_stack *stack, long index);
 size_t	stack_get_target_index(t_stack *s, uint value);
 bool	stack_is_sorted(t_stack *stack);
